style: fix 38 gate issues from the 20260826 scan
- G.FMT.03: merge io/jakarta and org/redis.clients import groups; keep
  within-group alphabetical order
- G.MET.01: parameter objects for >5-param methods (NodeFields/
  EdgeFields/RoundOpen/Delegation records)
- G.CON.10/12: writer runs on a managed single-thread pool, cooperative
  stop without Thread.interrupt
- G.CTL.02: trailing else branches; G.MET.03: no param reassignment
- G.ERR.02: test declares throws Throwable instead of catching it
- G.CMT.03: javadoc updated for the added audit/carrier params
- G.OTH.03: unused imports removed; G.FMT.04: fields back in field
  blocks; G.FMT.08: 4-space continuation indents; G.FMT.10/13/20:
  line length, wildcard spacing, 0L suffixes
- G.EXP.03: ternary arms unified to long; G.TYP.08/09: Locale.ROOT /
  UTF-8 explicit

Remaining 4 x G.MET.06 hits are Optional-returning method signatures
with no null returns (private Optional<T> helpers) — assessed as
scanner false positives pending team confirmation.

Tests: module 324 green

+++ common/agent-runtime-ext-java/agent-service-adapters/agent-service-adapters-agentcore-ext/src/main/java/com/openjiuwen/service/adapters/agentcore/ext/middleware/trajectory/TrajectoryLinkProperties.java
@@ -98,7 +98,7 @@ public class TrajectoryLinkProperties {
      * @return effective ttl seconds
      */
     public long effectiveTtlSeconds() {
-        return ttlSeconds != null && ttlSeconds > 0 ? ttlSeconds : DEFAULT_TTL_SECONDS;
+        return ttlSeconds != null && ttlSeconds.longValue() > 0L ? ttlSeconds.longValue() : DEFAULT_TTL_SECONDS;
     }
 
     /**
